Using 32bit variables for lengths - 16bit to small

+++ app/main_receive.c
@@ -16,25 +16,25 @@ uint32_t getCurrentTime()
 	return seconds * 1000;
 }
 
-uint16_t interfaceRead(uint8_t* buff, uint16_t maxLen)
+uint32_t interfaceRead(uint8_t* buff, uint32_t maxLen)
 {
 	usleep(1000);
 	return read(fileReadBuffer, buff, maxLen);
 }
 
-uint16_t interfaceWrite(uint8_t* buff, uint16_t len)
+uint32_t interfaceWrite(uint8_t* buff, uint32_t len)
 {
-	uint16_t retval = write(fileWriteBuffer, buff, len);
+	uint32_t retval = write(fileWriteBuffer, buff, len);
 	return retval;
 }
 
-uint16_t fileWrite(char* fileName, uint8_t* buff, uint16_t len)
+uint32_t fileWrite(char* fileName, uint8_t* buff, uint32_t len)
 {
 	FILE* output = fopen(fileName, "wb");
 	if (output == NULL)
 		return 0;
 
-	uint16_t retval = fwrite(buff, 1, len, output);
+	uint32_t retval = fwrite(buff, 1, len, output);
 	fclose(output);
 	return retval;
 }
